+++ passkit/lib/src/pkpass/pk_pass_image.dart
@@ -1,14 +1,14 @@
 import 'dart:typed_data';
 
-class PkPassImage {
-  PkPassImage({this.image1, this.image2, this.image3, this.localizedImages});
+class PkImage {
+  PkImage({this.image1, this.image2, this.image3, this.localizedImages});
 
   final Uint8List? image1;
   final Uint8List? image2;
   final Uint8List? image3;
   final Map<String, Map<int, Uint8List>>? localizedImages;
 
-  static PkPassImage? fromImages({
+  static PkImage? fromImages({
     Uint8List? image1,
     Uint8List? image2,
     Uint8List? image3,
@@ -17,7 +17,7 @@ class PkPassImage {
     if (image1 == null && image2 == null && image3 == null) {
       return null;
     }
-    return PkPassImage(image1: image1, image2: image2, image3: image3);
+    return PkImage(image1: image1, image2: image2, image3: image3);
   }
 
   /// [locale] is defined according to https://developer.apple.com/documentation/xcode/choosing-localization-regions-and-scripts